+++ solutions/devsprint-cecilia-saraiva-2/FinanceApp/Service/Entities/AccountProfile.swift
@@ -1,0 +1,27 @@
+//
+//  AccountProfile.swift
+//  FinanceApp
+//
+//  Created by Henrique Augusto on 16/08/22.
+//
+
+import Foundation
+
+struct Account: Decodable {
+
+    let agency: String
+    let account: String
+}
+
+extension Account {
+
+    func formattedAgency() -> String {
+
+        return "Agency: \(agency)"
+    }
+
+    func formattedAccount() -> String {
+
+        return "Account: \(account)"
+    }
+}
