+++ solutions/devsprint-cecilia-saraiva-2/FinanceApp/Service/Entities/UserProfile.swift
@@ -15,22 +15,3 @@ struct UserProfile: Decodable {
     let address: String
     let account: Account
 }
-
-struct Account: Decodable {
-
-    let agency: String
-    let account: String
-}
-
-extension Account {
-
-    func formattedAgency() -> String {
-
-        return "Agency: \(agency)"
-    }
-
-    func formattedAccount() -> String {
-
-        return "Account: \(account)"
-    }
-}
